fix(client): fx-desktop-v2 broker halts before signup confirmation poll.
The FxDesktopV2 auth broker was used as the superclass to both the
firstrun and Fennec auth brokers, when all these brokers really
needed was a Sync auth broker that communicated via WebChannels.
When the FxDesktopV2 started to be used as a standalone broker,
its default behavior of doing the confirmation poll caused
incorrect behavior if the user signed up for Sync from
`about:preferences#sync` - the original tab would redirect to
`/signup_complete` briefly before the browser re-displayed
`about:preferences#sync`.

This work extracts an FxSyncWebChannel authentication broker that
defines no behaviors. FxDesktopV2, FxFennecV1 and Firstrun all
extend from this new module and set any expected behaviors.
FxDesktopV2 now halts before the signup confirmation poll.

fixes #3330
