+++ app/scripts/models/auth_brokers/fx-desktop-v2.js
@@ -14,15 +14,14 @@ define(function (require, exports, module) {
   'use strict';
 
   var _ = require('underscore');
-  var Constants = require('lib/constants');
-  var FxSyncAuthenticationBroker = require('./fx-sync');
+  var FxSyncWebChannelAuthenticationBroker = require('./fx-sync-web-channel');
+  var HaltBehavior = require('views/behaviors/halt');
   var NavigateBehavior = require('views/behaviors/navigate');
   var p = require('lib/promise');
-  var WebChannel = require('lib/channels/web');
 
-  var proto = FxSyncAuthenticationBroker.prototype;
+  var proto = FxSyncWebChannelAuthenticationBroker.prototype;
 
-  var FxDesktopV2AuthenticationBroker = FxSyncAuthenticationBroker.extend({
+  var FxDesktopV2AuthenticationBroker = FxSyncWebChannelAuthenticationBroker.extend({
     afterSignUp: function (account) {
       var self = this;
       return p().then(function () {
@@ -35,6 +34,13 @@ define(function (require, exports, module) {
         }
       });
     },
+
+    defaultBehaviors: _.extend({}, proto.defaultBehaviors, {
+      // the browser is already polling, no need for the content server
+      // code to poll as well, otherwise two sets of polls are going on
+      // for the same user.
+      beforeSignUpConfirmationPoll: new HaltBehavior()
+    }),
 
     defaultCapabilities: _.extend({}, proto.defaultCapabilities, {
       chooseWhatToSyncCheckbox: false,
@@ -50,24 +56,7 @@ define(function (require, exports, module) {
       }
     }),
 
-    type: 'fx-desktop-v2',
-
-    commands: {
-      CAN_LINK_ACCOUNT: 'fxaccounts:can_link_account',
-      CHANGE_PASSWORD: 'fxaccounts:change_password',
-      DELETE_ACCOUNT: 'fxaccounts:delete_account',
-      LOADED: 'fxaccounts:loaded',
-      LOGIN: 'fxaccounts:login'
-    },
-
-    createChannel: function () {
-      var channel = new WebChannel(Constants.ACCOUNT_UPDATES_WEBCHANNEL_ID);
-      channel.initialize({
-        window: this.window
-      });
-
-      return channel;
-    }
+    type: 'fx-desktop-v2'
   });
 
   module.exports = FxDesktopV2AuthenticationBroker;
